pyproject.toml
[project]
name = "repl-nix-pharmainventorymanager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.40.2",
]
